+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Regression test instructions
 * For s3 upload, upload zip file and png only (no TIF files).
 * Better error handling of empty dataframes.
+* Fix the CLI entrypoint error in `run_sas_prep` due to missing parameter option for `prior_dist_s1_product` (see Issue #152: https://github.com/opera-adt/dist-s1/issues/152)
+
 
 ## [2.0.5] - 2025-08-15